[GameServer] 비동기 안정성 강화 및 필드 시스템 고도화
- GameSession::OnDisconnected() 내 널 포인터 체크 추가로 서버 크래시 방지
- Field 클래스의 모든 DoAsync 작업에서 shared_from_this()를 캡처하여 댕글링 포인터 위험 제거
- Field::EnterPlayer 매개변수를 weak_ptr로 변경하여 객체 소멸 시 대응 로직 추가
- FieldData 기반의 플레이어 스폰 위치 세팅 적용 (하드코딩 제거)
- 필드 파괴 예약 시 DestroyToken을 활용하여 유저 재입장 시 파괴 취소 로직 구현
- NavMesh 길찾기 좌표 변환 로직 정교화 (cm <-> m 변환 및 축 매핑 수정)
- SC_START_FIELD_LOADING 패킷에서 불필요한 start_pos 필드 제거
- CMake에 리소스 파일(JSON/NavMesh) 자동 복사 로직 추가 및 신규 테스트 맵 데이터 추가

+++ Server/GameServer/Contents/Field/FieldManager.h
@@ -13,7 +13,7 @@ public:
     void Init();
     void Load();
 
-    void Create(uint64 fieldId);
+    void Create(int32 fieldId);
     void Destroy(uint64 fieldId);
 
     shared_ptr<class Field> GetField(uint16 fieldId);
